Cinterop test works
