+++ performance/numerical/src/main/kotlin-native/launcher.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2010-2019 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the licenses/LICENSE.txt file.
+ */
+
+import org.jetbrains.benchmarksLauncher.*
+
+class NumericalLauncher : Launcher() {
+    override val benchmarks = BenchmarksCollection(
+            mutableMapOf(
+                    "Clang" to BenchmarkEntry(::clangBellardPi),
+                    "Konan" to BenchmarkEntry(::konanBellardPi)
+            )
+    )
+}
+
+fun clangBellardPi() {
+    for (n in 1 .. 1000 step 9)
+            cinterop.pi_nth_digit(n)
+}
+
+fun konanBellardPi() {
+    for (n in 1 .. 1000 step 9)
+            pi_nth_digit(n)
+}
