+++ performance/numerical/src/nativeInterop/cinterop/cinterop.def
@@ -1,0 +1,1 @@
+package = cinterop
