--- performance/numerical/src/nativeInterop/cinterop/pi.def
@@ -1,1 +1,0 @@
-package = pi
